Send timing information in batches once per minute
Sets up a thread with a queue interface to handle the upload of timing
information to the backend.

Previously this was done in the emitter thread after a skill or
subsystem had delivered their content to avoid slowing down the system.
However massively parallell calls like the "mycroft.stop" message would
starve the threadpool casuing the upload time to limit the execution
speed of these.